+++ src/components/FolderList.js
@@ -14,6 +14,7 @@ const StyledMenu = styled.div`
     border: 1px solid var(--color-primary);
     padding: 8px 12px;
     border-radius: 5px;
+    cursor: pointer;
   }
   .add-folder {
     color: var(--color-primary);
@@ -26,6 +27,7 @@ function FolderList({ data }) {
   return (
     <StyledMenu>
         <ul>
+          <li>전체</li>
           {data.map((list) => <li key={list.id}>{list.name}</li>)}
         </ul>
       <div className="add-folder">폴더 추가 +</div>
