+++ src/components/Pages/SharedPage/SharedPage.css
